+++ src/server/mod.rs
@@ -11,9 +11,9 @@ pub struct ServerConfig {
     dbname: String,
 
     // maximum number of robots to spawn
-    max_bots: u8,
+    max_bots: usize,
     // how quickly to regen robots
-    regen_rate: u8,
+    regen_rate: usize,
 } 
 
 pub fn get_config() -> ServerConfig {
@@ -56,10 +56,10 @@ pub fn get_config() -> ServerConfig {
         let dbname = matches.value_of("db").unwrap_or("ares").to_string();
 
         let max_bots = matches.value_of("max_bots").unwrap_or("10");
-        let max_bots = max_bots.parse::<u8>().expect("Could not parse max bots");
+        let max_bots = max_bots.parse::<usize>().expect("Could not parse max bots");
 
         let regen_rate = matches.value_of("regen_rate").unwrap_or("1");
-        let regen_rate = regen_rate.parse::<u8>().expect("Could not parse regen rate");
+        let regen_rate = regen_rate.parse::<usize>().expect("Could not parse regen rate");
 
         ServerConfig {
             dbuser, dbpw, dbhost, dbname, max_bots, regen_rate,
